+++ lexicmap/cmd/root.go
@@ -1,4 +1,4 @@
-// Copyright © 2023-2024 Wei Shen <shenwei356@gmail.com>
+// Copyright © 2023-2025 Wei Shen <shenwei356@gmail.com>
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -38,6 +38,7 @@ var RootCmd = &cobra.Command{
     Version: v%s
   Documents: https://bioinf.shenwei.me/LexicMap
 Source code: https://github.com/shenwei356/LexicMap
+Please cite: https://doi.org/10.1038/s41587-025-02812-8 Nature Biotechnology (2025)
 
 `, VERSION),
 }